requirements.txt
sentencepiece
transformers
torch
fastapi
starlette
aiofiles
python-multipart
Pillow
uvicorn
flask
multiprocessing
numpy
cv2